- Deleted DeleteRecord   - it was calling one method that can be called from Analyster where it should be, because that is were the file is deleted from.   - It should be a direct method of the Analyster window
